Move service and extensions down to shared analyzer layer
