+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/Core/Services/Precedence/PrecedenceKind.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.CodeAnalysis.LanguageServices
+namespace Microsoft.CodeAnalysis.Precedence
 {
     internal enum PrecedenceKind
     {
